Add ODCA Generator plugin to create site from API
The file is documented pretty well with comments, so I'll be brief here.
There are three parts to the "ODCA Plugin" -

1. An API "Client" which has all the knowledge of the structure of the
API.

2. A subclass of Jekyll::Page for each type of content, so that we can
apply custom logic to create page URLs and titles.

3. A Jekyll "Generator" which adds all the pages to the site.

The most unfortunate version of this approach is that it means we won't
be able to use Github Pages anymore :( -- custom plugins are not
supported in the "safe mode" that they use.