Rename SF locality_id "san-francisco" -> "sf"
We now have some logic relying on the structure of election_name fields
as "[locality]-[election year]", e.g. "sf-2018". Thus, locality IDs
can't easily contain dashes or else we won't be able to split the
election_name into locality reliably. This is kind of unfortunate since
"san-francisco" is better in URLs, so we should work to move back to
allowing dashes.