Use Web Components with React
Use Web Components to instantiate React components from Jekyll. This uses the
document-register-element polyfill for the Custom Elements. We create a HOC shim
for React to specify the name of the Custom Element. This allows the React
component to be used within React or instantiated from Jekyll by using the
Custom Element.