Include missing supporting/opposing committees
We calculate the supporting/opposing committees based on expenditures. So if the
committee hasn't made any expenditures, they won't show up in the list. This can
be confusing when contributions have been made, but no expenditures. In that
case, the totals for contributions appear, but it's unclear to which committees
those contributions were made.