Add link to Open Disclosure Alerts on homepage
In talking with Suzanne, we are hoping to pilot the alert emails more
broadly. Perhaps it is a bit premature to put a link to sign up for the
alerts on the OpenDisclosure.io homepage (I'll check with Suzanne), but
I'd like to have this ready to go in case we are ready.

Pretty straightforward layout, this adds an adjacent sibiling selector
to separate homepage description blocks with a border if there are
multiple.